refactor(editor): Extract remaining runData code into features/ndv/runData (no-changelog)
