+++ packages/frontend/editor-ui/src/features/ndv/runData/components/RunData.test.ts
@@ -18,7 +18,7 @@ import { waitFor } from '@testing-library/vue';
 import type { INodeExecutionData, ITaskData, ITaskMetadata } from 'n8n-workflow';
 import { setActivePinia } from 'pinia';
 import { useNodeTypesStore } from '@/stores/nodeTypes.store';
-import { useSchemaPreviewStore } from '@/stores/schemaPreview.store';
+import { useSchemaPreviewStore } from '@/features/ndv/runData/schemaPreview.store';
 import { useNDVStore } from '@/features/ndv/shared/ndv.store';
 
 const MOCK_EXECUTION_URL = 'execution.url/123';
